Conflicts:
	CMakeLists.txt
	agents/folderarchiveagent.desktop
	examples/mailreader/kmail-plasma/plasma-applet-kpapplet.desktop
	kleopatra/mainwindow_desktop.cpp
	kleopatra/systrayicon.cpp
	kmail/identity/identityaddvcarddialog.cpp
	kmail/tests/CMakeLists.txt
	kontact/plugins/planner/kcmplanner.desktop
	ktimetracker/ktimetracker_config_behavior.desktop
	ktimetracker/ktimetracker_config_display.desktop
	ktimetracker/ktimetracker_config_storage.desktop
	mobile/api/mail/CMakeLists.txt
	mobile/api/mail/composer.cpp

+++ identity/identityeditvcarddialog.cpp
@@ -58,21 +58,26 @@ IdentityEditVcardDialog::~IdentityEditVcardDialog()
 
 void IdentityEditVcardDialog::slotDeleteCurrentVCard()
 {
+    if (mVcardFileName.isEmpty())
+        return;
     if (KMessageBox::Yes == KMessageBox::questionYesNo(this, i18n("Are you sure to want to delete this vCard?"), i18n("Delete vCard"))) {
-        deleteCurrentVcard();
+        deleteCurrentVcard(true);
         reject();
     }
 }
 
-void IdentityEditVcardDialog::deleteCurrentVcard()
+void IdentityEditVcardDialog::deleteCurrentVcard(bool deleteOnDisk)
 {
     if (!mVcardFileName.isEmpty()) {
-        QFile file(mVcardFileName);
-        if (file.exists()) {
-            if (!file.remove()) {
-                KMessageBox::error(this, i18n("We cannot delete vCard file."), i18n("Delete vCard"));
+        if (deleteOnDisk) {
+            QFile file(mVcardFileName);
+            if (file.exists()) {
+                if (!file.remove()) {
+                    KMessageBox::error(this, i18n("We cannot delete vCard file."), i18n("Delete vCard"));
+                }
             }
         }
+        Q_EMIT vcardRemoved();
     }
 }
